Feat: reunindo operacoes
