--- databaseProject/src/main/java/com/klvth/databaseproject/DatabaseProject.java
@@ -1,16 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- */
-
-package com.klvth.databaseproject;
-
-/**
- *
- * @author Aluno
- */
-public class DatabaseProject {
-
-    public static void main(String[] args) {
-        System.out.println("Hello World!");
-    }
-}
